requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.18.0
numpy>=1.19.5
opencv-python>=4.5.4
scipy>=1.7.1
scikit-learn>=1.0.0
matplotlib>=3.5.0
pillow>=8.3.2
pydicom>=2.2.2
tqdm>=4.62.3
tensorboard>=2.7.0
einops>=0.4.1
timm>=0.5.4
pandas>=1.3.3
